+++ app/build.gradle
@@ -36,6 +36,7 @@ dependencies {
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
     implementation "io.reactivex.rxjava2:rxjava:2.1.9"
+    implementation "android.arch.lifecycle:extensions:1.1.1"
     implementation 'io.reactivex.rxjava2:rxandroid:2.0.1'
     implementation project(':openCVLibrary320')
 }
