+++ app/src/main/AndroidManifest.xml
@@ -5,7 +5,6 @@
 
     <uses-permission android:name="android.permission.CAMERA" />
     <uses-permission android:name="android.permission.SYSTEM_ALERT_WINDOW" />
-    <uses-permission android:name="android.permission.SYSTEM_ALERT_WINDOW" />
     <uses-permission android:name="android.permission.WRITE_SETTINGS"/>
 
     <uses-permission android:name="android.permission.ACTION_MANAGE_OVERLAY_PERMISSION" />
@@ -23,13 +22,13 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".Test2Activity"/>
+        <activity android:name=".MainActivity"/>
         <activity android:name=".ImageManipulationsActivity"/>
         <activity android:name=".Tutorial3Activity"/>
         <activity android:name=".OpenVcActivity"/>
         <activity android:name=".TestActivity"/>
 
-        <activity android:name=".MainActivity">
+        <activity android:name=".Test2Activity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
