+++ app/src/main/java/com/subject/opencvdemo/Service/FloatViewService.java
@@ -117,6 +117,14 @@ public class FloatViewService extends Service {
     }
 
     @Override
+    public void onRebind(Intent intent) {
+        super.onRebind(intent);
+
+
+
+    }
+
+    @Override
     public void onDestroy() {
         super.onDestroy();
         if (mFloatLayout != null) {
