+++ app/src/main/res/layout/floating_view.xml
@@ -3,6 +3,7 @@
     android:id="@+id/float_button"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
+    android:gravity="center"
     android:background="#000000"
     android:orientation="vertical">
